+++ lib/commanded/event_store/adapters/in_memory.ex
@@ -26,7 +26,8 @@ defmodule Commanded.EventStore.Adapters.InMemory do
   alias Commanded.EventStore.{EventData, RecordedEvent, SnapshotData}
 
   def start_link(opts \\ []) do
-    {start_opts, in_memory_opts} = Keyword.split(opts, [:name, :timeout, :debug, :spawn_opt])
+    {start_opts, in_memory_opts} =
+      Keyword.split(opts, [:debug, :name, :timeout, :spawn_opt, :hibernate_after])
 
     state = %State{
       name: Keyword.fetch!(opts, :name),
@@ -85,7 +86,7 @@ defmodule Commanded.EventStore.Adapters.InMemory do
   end
 
   @impl Commanded.EventStore.Adapter
-  def subscribe_to(adapter_meta, stream_uuid, subscription_name, subscriber, start_from) do
+  def subscribe_to(adapter_meta, stream_uuid, subscription_name, subscriber, start_from, _options) do
     event_store = event_store_name(adapter_meta)
 
     subscription = %Subscription{