+++ test/event/event_handler_error_handling_test.exs
@@ -4,7 +4,6 @@ defmodule Commanded.Event.EventHandlerErrorHandlingTest do
   import ExUnit.CaptureLog
 
   alias Commanded.DefaultApp
-  alias Commanded.Event.FailureContext
 
   alias Commanded.Event.ErrorAggregate.Events.{
     ErrorEvent,
@@ -13,12 +12,14 @@ defmodule Commanded.Event.EventHandlerErrorHandlingTest do
   }
 
   alias Commanded.Event.ErrorEventHandler
+  alias Commanded.Event.FailureContext
+  alias Commanded.Event.Handler
   alias Commanded.Helpers.EventFactory
 
   setup do
     start_supervised!(DefaultApp)
 
-    {:ok, handler} = ErrorEventHandler.start_link()
+    handler = start_supervised!(ErrorEventHandler)
 
     true = Process.unlink(handler)
 
@@ -111,11 +112,36 @@ defmodule Commanded.Event.EventHandlerErrorHandlingTest do
 
   test "should retry event handler after delay on error",
        %{handler: handler, ref: ref} do
-    send_error_event(handler, strategy: "retry", delay: 10)
+    send_error_event(handler, strategy: "retry", delay: 1)
 
-    assert_receive {:error, :failed, %{failures: 1, delay: 10}}
-    assert_receive {:error, :failed, %{failures: 2, delay: 10}}
-    assert_receive {:error, :too_many_failures, %{failures: 3, delay: 10}}
+    assert_receive {:error, :failed, %{failures: 1, delay: 1}}
+    assert_receive {:error, :failed, %{failures: 2, delay: 1}}
+    assert_receive {:error, :too_many_failures, %{failures: 3, delay: 1}}
+
+    assert_receive {:DOWN, ^ref, :process, ^handler, :too_many_failures}
+    refute Process.alive?(handler)
+  end
+
+  test "should retry event handler with `FailureContext` on error", %{handler: handler, ref: ref} do
+    send_error_event(handler, strategy: "retry_failure_context")
+
+    assert_receive {:error, :failed, %FailureContext{context: %{failures: 1}}}
+    assert_receive {:error, :failed, %FailureContext{context: %{failures: 2}}}
+    assert_receive {:error, :too_many_failures, %FailureContext{context: %{failures: 3}}}
+
+    assert_receive {:DOWN, ^ref, :process, ^handler, :too_many_failures}
+    refute Process.alive?(handler)
+  end
+
+  test "should retry event handler with `FailureContext` after delay on error",
+       %{handler: handler, ref: ref} do
+    send_error_event(handler, strategy: "retry_failure_context", delay: 1)
+
+    assert_receive {:error, :failed, %FailureContext{context: %{failures: 1, delay: 1}}}
+    assert_receive {:error, :failed, %FailureContext{context: %{failures: 2, delay: 1}}}
+
+    assert_receive {:error, :too_many_failures,
+                    %FailureContext{context: %{failures: 3, delay: 1}}}
 
     assert_receive {:DOWN, ^ref, :process, ^handler, :too_many_failures}
     refute Process.alive?(handler)
@@ -131,7 +157,8 @@ defmodule Commanded.Event.EventHandlerErrorHandlingTest do
     assert Process.alive?(handler)
 
     # Should ack errored event
-    assert GenServer.call(handler, :last_seen_event) == 1
+    %Handler{last_seen_event: last_seen_event} = :sys.get_state(handler)
+    assert last_seen_event == 1
   end
 
   defp send_error_event(handler, opts \\ []) do